Anoncoin Core v0.9.7.1-gost release

+++ src/db.cpp
@@ -44,7 +44,7 @@ void CDBEnv::EnvShutdown()
         LogPrintf("CDBEnv::EnvShutdown: Error %d shutting down database environment: %s\n", ret, DbEnv::strerror(ret));
     if (!fMockDb)
 #ifndef USING_PRE_HISTORIC_COMPILER
-        DbEnv(0).remove(strPath.c_str(), 0);
+        DbEnv((u_int32_t)0).remove(strPath.c_str(), 0);
 #else
        DbEnv((u_int32_t)0).remove(strPath.c_str(), 0);
 #endif
